Update: integrate libply written by Ares Lagae as an independent component in io
libply is written by Ares Lagae who agreed to share it with us under BSD license.
It is part of pcl yet can be used in standalone mode.
The built library is named libply but is refered as pcl_io_ply in CMakeLists.txt 
files to avoid potential conflicts.
The headers are located in pcl/io/ply if you need them.




git-svn-id: http://svn.pointclouds.org/pcl/trunk@4018 a9d63959-f2ad-4865-b262-bf0e56cfafb6